Corrected reference to USERDATA in path for scripts
Solves #449

The original referenced USERDATA/automation/jsr233 then gave the path to /etc/openhab2/automation/jsr233. According to Kai, the path is correct and the reference to USERDATA is incorrect. I updated the reference with CONF.

I also added a line that the Experimental Rules Engine is required to use JSR233. The docs do not mention this and it isn't until one clicks through to the old wiki that this information is presented.